% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.abundance main function.R
\name{gibbs.LDA}
\alias{gibbs.LDA}
\title{Gibbs sampling LDA without covariates}
\usage{
gibbs.LDA(y, ncomm, ngibbs, nburn, psi, gamma)
}
\arguments{
\item{y}{Matrix \eqn{L x S} with the quantities of each category in each sample unit. In text analysis we call this matrix bag of words.}

\item{ncomm}{Integer. Number of clusters (communities, topics, etc.)}

\item{ngibbs}{Integer. Total sample size}

\item{nburn}{Integer. Beginning of the sample to be thrown away}

\item{psi}{Numeric. Positive real number corresponding to the hyper parameter of the prior of \eqn{\phi}.}

\item{gamma}{Numeric. Positive real number corresponding to the hyper parameter of the prior of \eqn{V}.}
}
\description{
This function generates a sample of the parameters of interest using Metropolis Hastings algorithm combined with Gibbs sampling.
}
