% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated dataset}
\format{
A list containing two matrices:

\describe{
\item{y}{A matrix with the quantity of categories for each sample unit.}
\item{xmat}{A matrix with 4 covariables for each sample unit.}
}
}
\usage{
sim_data
}
\description{
A simulated dataset with 1000 sample units, 100 categories and 4 clusters.
}
\keyword{datasets}
